OPL-54: Moved BaseService and BaseModelValidation to core module
