+++ invoice/services/__init__.py
@@ -1,4 +1,3 @@
-from invoice.services.base import BaseService
 from invoice.services.invoice import InvoiceService
 from invoice.services.invoiceLineItem import InvoiceLineItemService
 from invoice.services.invoicePayments import InvoicePaymentsService
