+++ invoice/validation/billLineItem.py
@@ -1,5 +1,5 @@
 from invoice.models import BillItem
-from invoice.validation.base import BaseModelValidation
+from core.validation import BaseModelValidation
 
 
 class BillLineItemModelValidation(BaseModelValidation):
